Add restart from last run

+++ src/archive-village.ts
@@ -1,9 +1,11 @@
+import { existsSync } from 'node:fs';
 import path from 'node:path';
 import type { Browser, Page } from 'puppeteer';
 
 import { archiveActivity } from './archive-activity.ts';
 import { goToPage } from './go-to-page.ts';
 import { exportHTML } from './html.ts';
+import { getActivityPathsFromArchived } from './read-archived.ts';
 
 interface ArchiveVillageArgs {
     browser: Browser;
@@ -54,22 +56,30 @@ interface ArchiveVillagePhaseArgs {
 }
 async function archiveVillagePhase({ page, phase, villageName, resources, visitedActivities, baseDir, schoolYear }: ArchiveVillagePhaseArgs) {
     const filename = `${villageName.toLowerCase().replace(/[\s-]+/g, '-')}-phase-${phase}.html`;
-    console.info(`Archiving village: ${filename}`);
-    await page.evaluate(
-        ({ phase }) => {
-            sessionStorage.setItem('selectedPhase', `${phase}`);
-        },
-        { phase },
-    );
-    await goToPage(page, 'https://1v.parlemonde.org/?nopagination');
-    const html = await page.content();
-    const activityPaths = await exportHTML({
-        filename: path.join(baseDir, filename),
-        indexFileName: filename,
-        baseUrl: `/${schoolYear}`,
-        html,
-        resources,
-    });
+    const fullpath = path.join(baseDir, filename);
+    const activityPaths: string[] = [];
+    if (existsSync(fullpath)) {
+        activityPaths.push(...(await getActivityPathsFromArchived(fullpath)));
+    } else {
+        console.info(`Archiving village: ${filename}`);
+        await page.evaluate(
+            ({ phase }) => {
+                sessionStorage.setItem('selectedPhase', `${phase}`);
+            },
+            { phase },
+        );
+        await goToPage(page, 'https://1v.parlemonde.org/?nopagination');
+        const html = await page.content();
+        activityPaths.push(
+            ...(await exportHTML({
+                filename: path.join(baseDir, filename),
+                indexFileName: filename,
+                baseUrl: `/${schoolYear}`,
+                html,
+                resources,
+            })),
+        );
+    }
     while (activityPaths.length > 0) {
         const activityPath = activityPaths.pop();
         if (!activityPath) {
